+++ src/units.f90
@@ -13,7 +13,7 @@ contains
       implicit none
       type(units)            :: ret
       type(CFG_t), intent(in):: cfg
-      integer   , intent(in) :: me
+      integer(int64)   , intent(in) :: me
 
       ret%length      = get_unit_conv("length",      cfg, me, .True.)
       ret%energy      = get_unit_conv("energy",      cfg, me, .True.)
@@ -25,12 +25,12 @@ contains
    function get_unit_conv(field_name, cfg, me, bcast) result(factor)
       implicit none
       character(len=*), intent(in)        :: field_name
-      integer   , intent(in)              :: me
+      integer(int64)   , intent(in)              :: me
       logical, optional                   :: bcast
       logical                             :: bcast_loc
       type(CFG_t)                         :: cfg
       real(dp)                             :: factor
-      integer                             :: ierr
+      integer(int64)                             :: ierr
       character(len=300)                  :: unit_name
 
       if(present(bcast)) then
